test(web): isolate agent-session mention streaming helper
Move hideIncompleteMacroXml out of the Solid TextPart module so the
unit test does not load the markdown editor graph.
