+++ apps/web/src/features/block-agent/component/parts/TextPart.tsx
@@ -2,20 +2,7 @@
 
 import { StaticMarkdown } from '@core/component/LexicalMarkdown/component/core/StaticMarkdown';
 import { channelTheme } from '@core/component/LexicalMarkdown/theme';
-
-/**
- * Hide a trailing unclosed Macro mention tag so streaming text does not
- * flash raw `<m-document-mention>` XML. Complete tags stay intact.
- */
-export function hideIncompleteMacroXml(text: string): string {
-  const open = text.lastIndexOf('<m-');
-  if (open === -1) return text;
-  const after = text.slice(open);
-  if (/^<m-[a-zA-Z0-9_-]+>[\s\S]*<\/m-[a-zA-Z0-9_-]+>/.test(after)) {
-    return text;
-  }
-  return text.slice(0, open);
-}
+import { hideIncompleteMacroXml } from './hideIncompleteMacroXml';
 
 export function TextPart(props: { text: string; inFlight?: boolean }) {
   const markdown = () =>
