+++ apps/web/src/features/block-agent/component/parts/hideIncompleteMacroXml.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import { hideIncompleteMacroXml } from './TextPart';
+import { hideIncompleteMacroXml } from './hideIncompleteMacroXml';
 
 describe('hideIncompleteMacroXml', () => {
   const complete =
